Reorganize SDK module structure
