+++ modns-sdk/src/types/safe.rs
